Fixes to IntPtr conversion casting
Former-commit-id: cd2c550250ec1703b668273381d30ab302ee2498

+++ Anathema/Source/Tools/Scanners/PointerScanner/IPointerScannerMVP.cs
@@ -161,10 +161,10 @@ namespace Anathema
 
         public void SetMaxPointerOffset(String MaxPointerOffset)
         {
-            if (!CheckSyntax.CanParseValue(typeof(UInt64), MaxPointerOffset))
+            if (!CheckSyntax.CanParseValue(typeof(Int32), MaxPointerOffset))
                 return;
 
-            Model.SetMaxPointerOffset(Conversions.ParseValue(typeof(UInt64), MaxPointerOffset));
+            Model.SetMaxPointerOffset(Conversions.ParseValue(typeof(Int32), MaxPointerOffset));
             return;
         }
